fix(packages/ui, apps/web): GradientAnimatedText에서 motion 분리
