+++ apps/web/src/app/create/_components/AnimatedTitle/AnimatedTitle.css.ts
@@ -1,0 +1,8 @@
+import { vars } from '@repo/theme';
+import { style } from '@vanilla-extract/css';
+
+export const titleStyle = style({
+  width: '100%',
+  padding: `${vars.space[24]} 0`,
+  textAlign: 'center',
+});
