+++ packages/ui/src/components/GradientAnimatedText/GradientAnimatedText.tsx
@@ -1,34 +1,11 @@
-import { motion, HTMLMotionProps, MotionProps } from 'motion/react';
 import * as styles from './GradientAnimatedText.css';
-import { ReactNode, forwardRef } from 'react';
+import { ReactNode, forwardRef, ComponentPropsWithoutRef } from 'react';
 import { Slot } from '@radix-ui/react-slot';
 
 export type GradientAnimatedTextProps = {
   children: ReactNode;
   asChild?: boolean;
-} & HTMLMotionProps<'h1'>;
-
-const animationOptions = {
-  initial: {
-    y: '35vh',
-    scale: 2,
-    x: '-50%',
-    left: '50%',
-    position: 'absolute',
-  },
-  animate: {
-    y: 0,
-    scale: 1,
-    x: 0,
-    left: 'auto',
-    position: 'relative',
-  },
-  transition: {
-    type: 'spring',
-    duration: 0.6,
-    bounce: 0.22,
-  },
-} as MotionProps;
+} & ComponentPropsWithoutRef<'h1'>;
 
 /**
  * 그라디언트 애니메이션이 적용된 텍스트 컴포넌트입니다.
@@ -50,26 +27,21 @@ export const GradientAnimatedText = forwardRef<
   if (asChild) {
     return (
       <Slot ref={ref}>
-        <motion.div
-          className={`${styles.gradientTitleStyle} ${className}`}
-          {...animationOptions}
-          {...props}
-        >
+        <div className={`${styles.gradientTitleStyle} ${className}`} {...props}>
           {children}
-        </motion.div>
+        </div>
       </Slot>
     );
   }
 
   return (
-    <motion.h1
+    <h1
       ref={ref}
       className={`${styles.gradientTitleStyle} ${className}`}
-      {...animationOptions}
       {...props}
     >
       {children}
-    </motion.h1>
+    </h1>
   );
 });
